Plotting PCA or OPLS loadings

Plotting PCA or OPLS loadings

plotload(model, X, ppm, shift = c(0, 10), pc = 1,
  type = c("Statistical reconstruction"), title = "")

Arguments

model
PCA or OPLS model generated via MetaboMate package functions.

X
Input matrix with rows representing spectra

ppm
ppm variable

shift
ppm region to visualise.

pc
index of principal component to visualise, set to 1 if input model is OPLS

type
Type of loadigs visualisation, either 'Statistical reconstruction' or 'Backscaled' (see Details).

title
Plot title.

Details

OPLS: If type='Statistical recostruction' the function calculates the covariance (y axis) and Pearson's correlation (colouring) of the predictive OPLS scores with each X variable (x axis is ppm varaible). If type='Backscaled' the OPLS loadings are backscaled with X feature standard deviations. Results are plotted over ppm, coloured according to OPLS model weights. Often, the latter method visualises model importance more robust due to the presence of false positive correlations. PCA: Function always calculates the statistical recostruction.

References

Cloared, O., et al. (2005). Evaluation of the Orthogonal Projection on Latent Structure Model Limitations Caused by Chemical Shift Variability and Improved Visualization of Biomarker Changes in 1H NMR Spectroscopic Metabonomic Studies. Analytical Chemistry 77.2, 517-26.

See also