Probabilistic quotient normalisation

Probabilistic quotient normalisation

pqn(X, reference.idx = NA, TArea = F, add.DilF = NULL)

Arguments

X
Input matrix where rows represent samples.

reference.idx
Indices of spectra in X used to calculate reference spectrum (see Details).

TArea
Logical indicating if total area normalisation should be applied first (see Details).

add.DilF
Character string for new variable for dilution factor (will be exported as global environment variable). Can be left NULL if dilution factor should not be exported.

Details

It is sometimes favourable not to use all spectra to calculate a dilution reference (e.g. QC samples should generally be excluded). Therefore, a vector of indices can be specified with the parameter reference.idx and the respective spectra are used to calculate the median spectrum as a dilution reference. The parameter reference.idx can also be a single index, then the respective spectrum is used as a reference. If it is set N/A, all spectra in X are used to calculate the dilution reference spectrum. to Total area normalisation is integral part of the probabilistic quotient normalisation algorithm (see References), however, this sometimes distorts the spectra, i.e. is not suitable for normalisation. The parameter TArea can be set to TRUE or FALSE, depending if total area normalisation should be applied or not.

References

Dieterle, F., et al. (2006), Probabilistic Quotient Normalization as Robust Method to Account for Dilution of Complex Biological Mixtures. Application in 1H NMR Metabonomics, Analytical Chemistry, 78.3, 4281-90.