Calculating spectral quality indices

Calculating spectral quality indices

spec.quality(X, ppm, ppm.noise = c(9.4, 9.5), plot = T)

Arguments

X
NMR matrix, rows represent spectra.

ppm
Matching ppm vector.

ppm.noise
region for noise estimation, must be signal free across all spectra.

plot
Logical indicating if a summary plot should be produced.

Value

Returned is a data frame with TSP line width, residual water signal, normalised baseline estimation (the lower the less baseline fluctuations), estimated signal to noise ratio.

References

Eilers, P.H.C. (2004), Parametric Time Warping, Analytical Chemistry, 76.2, 404–411.

Bloemberg, T.G., et al. (2010), Improved parametric time warping for Proteomics, Chemometrics and Intelligent Laboratory Systems, 104.1, 65-74.

See also